pyproject.toml
[tool.poetry]
name = "blgit"
version = "0.0.0"
description = ""
authors = ["Gilad Kutiel <gilad.kutiel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jinja2 = "^3.1.4"
frontmatter = "^3.0.8"
markdown = "^3.7"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
